==> 02B_analysis_fecraw_markdown.Rmd <==
---
title: "GA Senate Runoffs"
author: "Aaron Kessler"
date: "`r format(Sys.time(), '%B %d, %Y')`"
output: 
  html_document:
    toc: true
    toc_float: true
    code_folding: hide
    # code_download: true
    df_print: kable
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(lubridate)
library(janitor)
library(plotly)
library(widgetframe)
library(scales)
library(zoo)
library(summarytools)
library(writexl)
library(tidycensus)
library(tigris)
library(sf)
library(leaflet)
library(tmap)
library(tmaptools)
library(RColorBrewer)
library(htmlwidgets)
library(htmltools)
library(widgetframe)
library(sp)
library(raster)
library(mapview)
library(scales)
library(viridis)
library(tidyverse)
options(tigris_class = "sf")
options(dplyr.summarise.inform = FALSE)


```


```{r dataload, include=FALSE, cache=TRUE}
#load data files created in step 01
actblue_contribs <- readRDS("processed_data/actblue_fecraw_contribs_postelex_GASEN.rds")
winred_contribs <- readRDS("processed_data/winred_fecraw_contribs_postelex_GASEN.rds")


#ensure that nothing slipped through during processing and we're just looking at 
#individual earmarked contributions
actblue_contribs <- actblue_contribs %>% 
  filter(
    entity_type == "IND", # only individual contributions
    is.na(memo_code), # remove memo lines
    contribution_purpose_descrip == "Earmark" # remove donations that are not for a specific candidate
  )

winred_contribs <- winred_contribs %>% 
  filter(
    entity_type == "IND", # only individual contributions
    is.na(memo_code), # remove memo lines
  )


```


```{r zipdataprep, include=FALSE}

# Prepare zipcode data ####

topzips_gop <- winred_contribs %>% 
  mutate(
    zip5 = str_sub(contributor_zip_code, 1, 5)
  ) %>% 
  group_by(zip5) %>% 
  summarise(sumtotal = sum(contribution_amount)) %>% 
  arrange(desc(sumtotal))

# topzips_gop


topzips_dem <- actblue_contribs %>% 
  mutate(
    zip5 = str_sub(contributor_zip_code, 1, 5)
  ) %>% 
  group_by(zip5) %>% 
  summarise(sumtotal = sum(contribution_amount)) %>% 
  arrange(desc(sumtotal))

# topzips_dem


### compare D and R in each zip ##

colnames(topzips_dem) <- c("zip_code", "demtotal")
colnames(topzips_gop) <- c("zip_code", "goptotal")

#join
zipcompare <- full_join(topzips_dem, topzips_gop)

#change nas to 0
zipcompare <- zipcompare %>% 
  replace(., is.na(.), 0)

#party with more
zipcompare <- zipcompare %>% 
  mutate(
    winner = ifelse(demtotal>goptotal, "D", "R"),
    advantage = abs(demtotal-goptotal)  
  )


### bring in the zipcode data table for adding names to zips 
zip_codes_lookup <- read_csv("zip_data/zip-codes-database-STANDARD.csv")

zip_codes_lookup <- zip_codes_lookup %>% 
  clean_names()

zip_codes_lookup <- zip_codes_lookup %>% 
  mutate(
    city = str_to_title(city),
    name = paste0(city, ", ", state),
    zipname = paste0(zip_code, " (", name, ")")
  ) %>% 
  select(zip_code, city, state, zipname)

#remove dups
zip_codes_lookup <- zip_codes_lookup %>% 
  distinct(zip_code, city, state, zipname)


#join to our table
zipcompare <- left_join(zipcompare, zip_codes_lookup)


### ADDING IN IRS ZIP CODE ECONOMIC DATA ##

irs_zip_raw <- read_csv("zip_data/16zpallagi.csv", 
                        col_types = cols(STATEFIPS = col_character(), 
                                         zipcode = col_character()))

irszips_allcols <- irs_zip_raw %>% 
  clean_names()

names(irszips_allcols)


#looking for any 4 digit zips by mistake?
irszips_allcols %>% 
  filter(str_length(zipcode)==4)

#fix 4-digit zips to add back preceding 0
zipscorrected <- if_else(str_length(irszips_allcols$zipcode)==4, paste0("0", irszips_allcols$zipcode), irszips_allcols$zipcode)
irszips_allcols$zipscorrected <- zipscorrected
#check cols
irszips_allcols %>% 
  select(zipcode, zipscorrected) %>% 
  filter(zipcode != zipscorrected)
#make replacement
irszips_allcols$zipcode <- irszips_allcols$zipscorrected


#select just AGI columns
irs_zips_agi <- irszips_allcols %>% 
  select(statefips, state, zipcode, agi_stub, total_returns = n1, agi_value = a00100)

#take out the state-wide totals, zip = 0. Also remove "other" designation, 99999
irs_zips_agi <- irs_zips_agi %>% 
  filter(zipcode != "0",
         zipcode != "99999")

head(irs_zips_agi)


#looking for any 4 digit zips by mistake?
irs_zips_agi %>% 
  filter(str_length(zipcode)==4)


### create column translating agi_stub to description, per documentation
# 1 = $1 under $25,000
# 2 = $25,000 under $50,000
# 3 = $50,000 under $75,000
# 4 = $75,000 under $100,000
# 5 = $100,000 under $200,000
# 6 = $200,000 or more

irs_zips_agi <- irs_zips_agi %>% 
  mutate(
    agi_stubcategory = case_when(
      agi_stub == 1 ~ "under $25K",
      agi_stub == 2 ~ "$25K to under $50K",
      agi_stub == 3 ~ "$50K to under $75K",
      agi_stub == 4 ~ "$75K to under $100K",
      agi_stub == 5 ~ "$100K to under $200K",
      agi_stub == 6 ~ "$200K or more",
      TRUE ~ "other"
    )
  )


#sum up all agi/returns for each zip
irs_zips_agi_grouped <- irs_zips_agi %>% 
  group_by(zipcode) %>% 
  summarise(num_returns = sum(total_returns), total_agi_value = sum(agi_value)) %>% 
  mutate(avg_agi = total_agi_value/num_returns*1000)

head(irs_zips_agi_grouped)

#rank by agi
irs_zips_agi_grouped <- irs_zips_agi_grouped %>% 
  mutate(avg_agi_rank = rank(desc(avg_agi))) %>% 
  arrange(avg_agi_rank)

irs_zips_agi_grouped

# join candidates by zip with processed IRS table
zipcompare <- left_join(zipcompare, irs_zips_agi_grouped, by = c("zip_code" = "zipcode"))


#save table as rds for later use
saveRDS(zipcompare, "zip_data/zipcompare.rds")
# save as excel for sharing
write_xlsx(zipcompare, "output/zipcompare.xlsx")


```


# Totals 

#### Actblue

```{r}

actblue_contribs %>% 
  group_by(ga_candidate) %>% 
  summarise(total_dollars = sum(contribution_amount), avg_amount = mean(contribution_amount), median_amount = median(contribution_amount)) 

```

#### WinRed
```{r}
winred_contribs %>% 
  group_by(ga_candidate) %>% 
  summarise(total_dollars = sum(contribution_amount), avg_amount = mean(contribution_amount), median_amount = median(contribution_amount)) 

```

###
# In-State vs. Out-of-State

### Party Totals 

**Dem overall totals**
```{r}

dem_inout <- actblue_contribs %>% 
  group_by(in_out_state) %>% 
  summarise(total_dollars = sum(contribution_amount)) %>% 
  mutate(
    pct_of_total = round_half_up(total_dollars/sum(total_dollars)*100, 1)
    ) %>% 
  ungroup()

dem_inout
```

**GOP overall totals**
```{r}

gop_inout <- winred_contribs %>%
  group_by(in_out_state) %>%
  summarise(total_dollars = sum(contribution_amount)) %>%
  mutate(
    pct_of_total = round_half_up(total_dollars/sum(total_dollars)*100, 1)
    ) %>% 
  ungroup()

gop_inout
```


### Dem Totals By State, Candidate

**Dem totals by top states**
```{r}

dem_bystate <- actblue_contribs %>% 
  group_by(contributor_state) %>% 
  summarise(total_dollars = sum(contribution_amount)) %>% 
  mutate(
    pct_of_total = round_half_up(total_dollars/sum(total_dollars)*100, 1)
    ) %>% 
  arrange(desc(total_dollars)) %>% 
  ungroup() %>% 
  head(15)

dem_bystate
```


**Dem candidate totals**
```{r}

dem_inout_bycand <- actblue_contribs %>% 
  group_by(ga_candidate, in_out_state) %>% 
  summarise(total_dollars = sum(contribution_amount)) %>% 
  mutate(
    pct_of_total = round_half_up(total_dollars/sum(total_dollars)*100, 1)
    ) %>% 
  ungroup()

dem_inout_bycand


```

```{r}

#chart it out
d <- ggplot(dem_inout, aes(in_out_state, total_dollars)) + geom_col(fill = "darkblue") + coord_flip() +
  theme_minimal()

#add extra elements to the chart and convert to ggplotly
d2 <- d + labs(title = "Dems: In vs. Out of State",
               subtitle = "(Since Nov. 4th)",
               caption = "Source: Bloomberg Analysis, FEC",
               x ="", y = "") +
  theme(plot.title = element_text(hjust = 0.5)) +
  theme(plot.subtitle = element_text(hjust = 0.5)) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_y_continuous(labels=dollar) 

d2





```





```{r}

# The palette:
cbPalette <- c("#508104","#db8200")

d <- ggplot(data = dem_inout_bycand, aes(x = ga_candidate, y = total_dollars, fill = in_out_state)) +
  geom_col(position = "dodge") + coord_flip() + theme_minimal()

d2 <- d + labs(title="Dems Candidates: In vs. Out of State",
               subtitle = "(Since Nov. 4th)",
               caption = "Source: Source: Bloomberg Analysis, FEC",
               x ="", y = "") +
  theme(plot.title = element_text(hjust = 0.5)) +
  theme(plot.subtitle = element_text(hjust = 0.5)) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_y_continuous(labels=dollar) +
  scale_fill_manual(values=cbPalette) +
  theme(legend.title=element_blank())

d2



```


### GOP Totals By State, Candidate

**GOP totals by top states**
```{r}

gop_bystate <- winred_contribs %>% 
  group_by(contributor_state) %>% 
  summarise(total_dollars = sum(contribution_amount)) %>% 
  mutate(
    pct_of_total = round_half_up(total_dollars/sum(total_dollars)*100, 1)
    ) %>% 
  arrange(desc(total_dollars)) %>% 
  ungroup() %>% 
  head(15)

gop_bystate
```



**GOP candidate totals**
```{r}

gop_inout_bycand <- winred_contribs %>%
  group_by(ga_candidate, in_out_state) %>%
  summarise(total_dollars = sum(contribution_amount)) %>%
  mutate(
    pct_of_total = round_half_up(total_dollars/sum(total_dollars)*100, 1)
    ) %>% 
  ungroup()

gop_inout_bycand


```


```{r}

#chart it out
d <- ggplot(gop_inout, aes(in_out_state, total_dollars)) + geom_col(fill = "darkred") + coord_flip() +
  theme_minimal()

#add extra elements to the chart and convert to ggplotly
d2 <- d + labs(title = "Republicans: In vs. Out of State",
               subtitle = "(Since Nov. 4th)",
               caption = "Source: Bloomberg Analysis, FEC",
               x ="", y = "") +
  theme(plot.title = element_text(hjust = 0.5)) +
  theme(plot.subtitle = element_text(hjust = 0.5)) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_y_continuous(labels=dollar)

d2





```





```{r}

# The palette:
cbPalette <- c("#508104","#db8200")

d <- ggplot(data = gop_inout_bycand, aes(x = ga_candidate, y = total_dollars, fill = in_out_state)) +
  geom_col(position = "dodge") + coord_flip() + theme_minimal()

d2 <- d + labs(title="Republican Candidates: In vs. Out of State",
               subtitle = "(Since Nov. 4th)",
               caption = "Source: Source: Bloomberg Analysis, FEC",
               x ="", y = "") +
  theme(plot.title = element_text(hjust = 0.5)) +
  theme(plot.subtitle = element_text(hjust = 0.5)) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_y_continuous(labels=dollar) +
  scale_fill_manual(values=cbPalette) +
  theme(legend.title=element_blank())

d2



```


# Zip Codes


**Top Zips for Democratic Total Contributions (ActBlue)**

```{r}

#pull out top 10 zips for the dems
dem10 <- zipcompare %>% 
  arrange(desc(demtotal)) %>% 
  head(15)

# dem10

#reorder factor to allow for descending bars
dem10 <- dem10 %>%
  mutate(zipname = fct_reorder(zipname, demtotal)) 

#export to file
write_csv(dem10, "output/dem10.csv")


#chart it out
d <- ggplot(dem10, aes(zipname, demtotal)) + geom_col(fill = "darkblue") + coord_flip() +
  theme_minimal()

#add extra elements to the chart and convert to ggplotly
d2 <- d + labs(title="GA Runoffs: Top ActBlue zip codes",
               # subtitle = "A subtitle",
               caption = "Source: FEC, Bloomberg Analysis",
               x ="", y = "") +
  theme(plot.title = element_text(hjust = 0.5)) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_y_continuous(labels=dollar) 

d2



# dd <- ggplotly(d2) 
# 
# dd_nomenu <- dd %>% config(displayModeBar = FALSE)
# dd_nomenu
# 
# #save as embeddable format
# # htmlwidgets::saveWidget(frameableWidget(dd), 'demtopzip_plt.html')
# htmlwidgets::saveWidget(frameableWidget(dd_nomenu), 'demtopzip_chart_nm.html')
# 
# #save as RDS object
# saveRDS(dd_nomenu, "zip_data/demtopzip_chart_nm.rds")
```


**Top Zips for Democratic Advantage (ActBlue)**

```{r}

#pull out top zips for the dems
dem10_adv <- zipcompare %>% 
  filter(winner == "D") %>% 
  arrange(desc(advantage)) %>% 
  head(15)

#reorder factor to allow for descending bars
dem10_adv <- dem10_adv %>%
  mutate(zipname = fct_reorder(zipname, advantage)) 

#export to file
write_csv(dem10_adv, "output/dem10_adv.csv")


#chart it out
d <- ggplot(dem10_adv, aes(zipname, advantage)) + geom_col(fill = "darkblue") + coord_flip() +
  theme_minimal()

#add extra elements to the chart and convert to ggplotly
d2 <- d + labs(title="GA Runoffs: Top Zips Where Grassroots Dems Outraised GOP",
               subtitle = "(bars show advantage in dollars raised)",
               caption = "Source: FEC, Bloomberg Analysis",
               x ="", y = "") +
  theme(plot.title = element_text(hjust = 0.5)) +
  theme(plot.subtitle = element_text(hjust = 0.5)) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_y_continuous(labels=dollar) 

d2

```




**Top Zips for Republican Total Contributions (WinRed)**

```{r}

#pull out top 10 zips for the gop
gop10 <- zipcompare %>% 
  arrange(desc(goptotal)) %>% 
  head(15)

# gop10

#reorder factor to allow for descending bars
gop10 <- gop10 %>%
  mutate(zipname = fct_reorder(zipname, goptotal)) 

#export to file
write_csv(gop10, "output/gop10.csv")


#chart it out
p <- ggplot(gop10, aes(zipname, goptotal)) + geom_col(fill = "darkred") + coord_flip() +
  theme_minimal()

#add titles and other extras
p2 <- p + labs(title="Ga Senate Runoffs: Top WinRed zip codes",
               # subtitle = "A subtitle",
               caption = "Source: FEC, Bloomberg Analysis",
               x ="", y = "") +
  theme(plot.title = element_text(hjust = 0.5)) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_y_continuous(labels=dollar) 

p2

# pp <- ggplotly(p2) 
# 
# pp
# 
# pp_nomenu <- pp %>% config(displayModeBar = FALSE)
# pp_nomenu
# 
# #save as embeddable format
# # htmlwidgets::saveWidget(frameableWidget(pp), 'goptopzip_plt.html')
# htmlwidgets::saveWidget(frameableWidget(pp_nomenu), 'goptopzip_plt_nm.html')

#save as RDS object
# saveRDS(pp_nomenu, "zip_data/goptopzip_plt_nm.rds")




```

**Top Zips for Republican Advantage (WinRed)**

```{r}


#pull out top zips 
gop10_adv <- zipcompare %>% 
  filter(winner == "R") %>% 
  arrange(desc(advantage)) %>% 
  head(15)

#reorder factor to allow for descending bars
gop10_adv <- gop10_adv %>%
  mutate(zipname = fct_reorder(zipname, advantage)) 

#export to file
write_csv(gop10_adv, "output/gop10_adv.csv")


#chart it out
d <- ggplot(gop10_adv, aes(zipname, advantage)) + geom_col(fill = "darkred") + coord_flip() +
  theme_minimal()

#add extra elements to the chart and convert to ggplotly
d2 <- d + labs(title="GA Runoffs: Top Zips Where Grassroots GOP Outraised Dems",
               subtitle = "(bars show advantage in dollars raised)",
               caption = "Source: FEC, Bloomberg Analysis",
               x ="", y = "") +
  theme(plot.title = element_text(hjust = 0.5)) +
  theme(plot.subtitle = element_text(hjust = 0.5)) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_y_continuous(labels=dollar) 

d2

```






## Zip Maps

```{r}
#bring in comparison table created above
zipcompare <- readRDS("zip_data/zipcompare.rds")

head(zipcompare)
names(zipcompare)

zipcompare$GEOID <- zipcompare$zip_code

#remove any negative values
zipcompare <- zipcompare %>% 
  filter(demtotal >= 0,
         goptotal >= 0)

zipcompare <- zipcompare %>% 
  filter(!str_detect(GEOID, "^99"),
         !str_detect(GEOID, "^96"),
         !str_detect(GEOID, "^006"),
         !str_detect(GEOID, "^007"),
         !str_detect(GEOID, "^009")
           )


#### get zip code lat/long points from census gazeteer file
zip_points <- read_csv("zip_data/zip_points.csv", 
                       col_types = cols(ALAND = col_skip(), 
                                        ALAND_SQMI = col_skip(), AWATER = col_skip(), 
                                        AWATER_SQMI = col_skip()))

colnames(zip_points) <- c("GEOID", "lat", "lon")

#join data
zip_map <- inner_join(zipcompare, zip_points)

zip_map$winner <- as.factor(zip_map$winner)

#add dollar formatting
zip_map$demdisplay <- dollar(zip_map$demtotal)
zip_map$gopdisplay <- dollar(zip_map$goptotal)
```

National zipcode map of Dem vs. GOP advantage in contributions
```{r}
#### MAPPING POINTS ##### ---------------------------------------

factpal <- colorFactor(c("blue","red"), zip_map$winner)

#labels
labs1 <- lapply(seq(nrow(zip_map)), function(i) {
  paste0( '<p>', 'Zip code: ', '<strong>', zip_map[i, "GEOID"], '</strong></p>',
          '<p></p>', 
          "Democrats: ", zip_map[i, "demdisplay"],
          '<p></p>', 
          "Republicans: ", zip_map[i, "gopdisplay"]
  ) 
})

m1 <- leaflet(zip_map) %>% 
  addTiles() %>%
  addCircles(lng = ~lon, lat = ~lat, weight = .4,
             stroke = FALSE, fillOpacity = .25,
             radius = ~sqrt(advantage) * 300, 
             fillColor = ~factpal(winner),
             label = lapply(labs1, HTML)
  ) %>%
  addControl("ActBlue vs. WinRed - GA Senate candidate contributions by zip code", position = "topright") 
# %>% 
#   setView(-96, 37.8, zoom=4) 

m1

#save to frameable file for CMS
htmlwidgets::saveWidget(frameableWidget(m1),'gasenate_actblue_vs_winred_zippoints.html')

# save as rds object
saveRDS(m1, "savedmap_zipcompare.rds")



```



Atlanta-area-only zipcode map of Dem vs. GOP advantage in contributions
```{r}

#filter for just GA
zip_map <- zip_map %>% 
  filter(state == "GA")

factpal <- colorFactor(c("blue","red"), zip_map$winner)

#labels
labs1 <- lapply(seq(nrow(zip_map)), function(i) {
  paste0( '<p>', 'Zip code: ', '<strong>', zip_map[i, "GEOID"], '</strong></p>',
          '<p></p>', 
          "Democrats: ", zip_map[i, "demdisplay"],
          '<p></p>', 
          "Republicans: ", zip_map[i, "gopdisplay"]
  ) 
})

m2 <- leaflet(zip_map) %>% 
  addTiles() %>%
  addCircles(lng = ~lon, lat = ~lat, weight = .4,
             stroke = FALSE, fillOpacity = .25,
             radius = ~sqrt(advantage) * 20, 
             fillColor = ~factpal(winner),
             label = lapply(labs1, HTML)
  ) %>%
  addControl("ActBlue vs. WinRed - GA Senate candidate contributions by zip code", position = "topright") %>%
  setView(-84.3880, 33.7490, zoom = 9)

m2

#save to frameable file for CMS
htmlwidgets::saveWidget(frameableWidget(m2),'gasenate_actblue_vs_winred_zippoints_ATLONLY.html')

# save as rds object
saveRDS(m2, "savedmap_zipcompare_ATLONLY.rds")



```




```{r}

#### top zips for each party

t_gop <- zipcompare %>% 
  select(zip_code, total = goptotal) %>%
  mutate(party = "R") %>% 
  arrange(desc(total)) %>% 
  head(250)

t_dem <- zipcompare %>% 
  select(zip_code, total = demtotal) %>% 
  mutate(party = "D") %>% 
  arrange(desc(total)) %>% 
  head(250)

#combine
t_both <- bind_rows(t_dem, t_gop)

t_both %>% write_xlsx("output/t_both.xlsx")

```


```{r}




```


```{r}




```
